requirements.txt
python-telegram-bot[job-queue]==20.3
gspread
oauth2client
requests
python-dotenv
pytest